bug-hunt #729: acpi_cpufreq/cpufreq - P-state transition safety (read-modify-write PERF_CTL, fix control width, fix stub)
Three interrelated bugs found and fixed:

1. cpupstate_set_state() and cpufreq_fast_switch() clobbered PERF_CTL
   upper bits (VID in bits 15:8, IDA/turbo in bits 32-33) by writing
   only the low byte. Fixed with read-modify-write that preserves all
   bits except the P-state number (bits 7:0). This prevents silent
   re-enabling of turbo after cpufreq_boost_disable() and ensures VID
   settings are preserved on older Intel Enhanced SpeedStep CPUs.

2. struct cpupstate_state.control/.status were uint8_t, truncating
   the 32-bit ACPI _PSS Control/Status DWORD values. Expanded to
   uint32_t so the full PERF_CTL encoding (FID, VID, platform bits)
   is preserved through the ACPI parsing path and available for the
   transition functions.

3. acpi_cpufreq_target() returned 0 (success) despite being a stub.
   Changed to return -1 so callers know the transition did not occur.

+++ src/include/cpupstate.h
@@ -27,8 +27,8 @@ struct cpupstate_state {
     uint32_t power;        /* mW */
     uint32_t transition_latency; /* us */
     uint32_t bus_master_latency; /* us (bus master latency) */
-    uint8_t  control;      /* Control value (written to MSR) */
-    uint8_t  status;       /* Status value */
+    uint32_t control;      /* Control value (written to PERF_CTL MSR) */
+    uint32_t status;       /* Status value (read from PERF_STATUS) */
 };
 
 /* CPU P-state controller */
